requirements.txt
# torch / torchvision / torchaudio は Dockerfile 側で cu118 変種を固定
faster-whisper[gpu]==1.0.3
pyannote.audio==3.1.1
python-dotenv

transformers
accelerate
punctuators==0.0.5
git+https://github.com/huggingface/diarizers.git

onnxruntime-gpu==1.17.1

numpy<2.0
packaging<25.0
tqdm
runpod
huggingface_hub
pyyaml